Added TransformComponentBase
Added TransformComponent to Entity

Changed all doubles back to float

+++ Common/Game/Interfaces/IManager.cs
@@ -44,7 +44,7 @@
     /// <param name="maxTime">
     ///   The max time the manager can spend on its update, in seconds.
     /// </param>
-    void Update(double deltaTime, double maxTime);
+    void Update(float deltaTime, float maxTime);
 
     /// <summary>
     ///   Shuts down the manager without releasing resources.
